`download()`: Clear libpath environment variables before invoking `curl`
We generally don't have a problem when invoking external executables,
however occasionally we are doing something such as running `rr`, which
needs `libgcc_s` from the Julia CSL distribution, and the only way to
get `libgcc_s` for `rr` is to tack its location on to the end of
`LD_LIBRARY_PATH`.  This causes all other executables to search that
directory as well, and right now, that directory is Julia's entire
private libdir, which includes many things such as `libcurl.so`.  This
breaks a system-provided `curl`.

This wil be fixed by JLL stdlib bundling, as we will have much more
fine-grained control over which libraries are added to the paths for
which executables, however in the meantime, an easy fix is to isolate
the system `curl` from our libraries by removing the libpath environment
variables before invoking it.